finished clash royale page
