+++ frontend/components/Footer/index.tsx
@@ -1,0 +1,9 @@
+import styles from "../../styles/Footer.module.css";
+
+export default function Footer() {
+  return (
+    <footer className={styles.footer}>
+      <h2>Copyright 2021</h2>
+    </footer>
+  );
+}
