+++ frontend/styles/Home.module.css
@@ -87,5 +87,5 @@
 
 .buttons button:hover {
   background-color: transparent;
-  color: white;
+  color: var(--white-color);
 }
